bot/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-dotenv==1.0.0
pinecone-client==3.0.0
openai>=1.6.1
langchain-openai==0.0.2.post1
pandas==2.0.3
jinja2==3.1.2
python-multipart==0.0.6
aiofiles==23.2.1